Task25_1 new_

+++ Task25_1/Program.cs
@@ -1,15 +1,15 @@
 ﻿double Degree(int a, int b)
 {
-   double resdeg = 1;
-   double res = 0;
-   for(int i=1;i<=b;i++)
+    double resdeg = 1;
+    double res = 0;
+    for (int i = 1; i <= b; i++)
     {
-         resdeg = resdeg*a;
+        resdeg = resdeg * a;
     }
-    if (b>0) res = resdeg;
-    else res = 1/resdeg;
-        // else res = resdeg;
-    
+    if (b > 0) res = resdeg;
+    else res = (double)1 / resdeg;
+    // else res = resdeg;
+
     return res;
 }
 
@@ -19,6 +19,6 @@ Console.WriteLine("Введите число B:");
 int y = Convert.ToInt32(Console.ReadLine());
 
 
-    double degree = Degree(x, y);
-    Console.WriteLine($"{degree}");
+double degree = Degree(x, y);
+Console.WriteLine($"{degree}");
